fix chart component

+++ src/app/topic/topic.component.ts
@@ -9,7 +9,6 @@ import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { DataService } from 'app/services/data.service';
 import { Topic } from '../models/topic';
-// import { Result } from '../models/result';
 import { AuthService } from 'app/services/auth.service';
 import { pageTransition } from '../animations';
 import { ToastrService } from 'app/services/toastr.service';
@@ -23,12 +22,20 @@ import { ToastrService } from 'app/services/toastr.service';
 export class TopicComponent implements OnInit {
   _id: string;
   topic: Topic;
-  // results: Result[];
-  totalVotes: number;
+  // totalVotes: number;
   hasVoted = false;
   currentUser: string;
   toggleButton = 'View Chart';
   view = 'data';
+
+  // computed values
+  get totalVotes() {
+    if (this.topic) {
+      return this.topic.results.map(result => result.votes)
+        .reduce((a, b) => a + b);
+    }
+    return 0;
+  }
 
   constructor(
     private _route: ActivatedRoute,
@@ -44,9 +51,10 @@ export class TopicComponent implements OnInit {
   ngOnInit(): void {
     this._data.fetchTopicById(this._id)
       .subscribe(topic => {
+        console.log(this.totalVotes);
         this.topic = topic;
-        // this.results = topic.results;
-        this.totalVotes = this.getTotalVotes();
+        console.log(this.totalVotes);
+        // this.totalVotes = this.getTotalVotes();
         this.currentUser = this._auth.isValidated();
         this.topic.voters.forEach(voter => {
           if (voter === this.currentUser) {
@@ -56,10 +64,10 @@ export class TopicComponent implements OnInit {
       });
   }
 
-  getTotalVotes(): number {
-    return this.topic.results.map(result => result.votes)
-      .reduce((a, b) => a + b);
-  }
+  // getTotalVotes(): number {
+  //   return this.topic.results.map(result => result.votes)
+  //     .reduce((a, b) => a + b);
+  // }
 
   castVote(res): void {
     if (!this.hasVoted) {
@@ -72,21 +80,22 @@ export class TopicComponent implements OnInit {
       // TODO - refactor UI feedback into Promise return
       this.topic.voters.push(this.currentUser);
       this.hasVoted = true;
-      this.totalVotes = this.getTotalVotes();
+      // this.totalVotes = this.getTotalVotes();
       this._data.castVote(this._id, this.topic);
       this._toastr.success('Thanks for your vote');
     }
   }
 
-  deleteTopic(id: string) {
-    this._data.deleteTopic(id).then(() => {
-      this._toastr.warning('Topic deleted!');
-      this._router.navigate(['/home']);
-    });
-
+  deleteTopic(id: string): void {
+    if (window.confirm('Are you sure you want to permanently delete this topic and its data?')) {
+      this._data.deleteTopic(id).then(() => {
+        this._toastr.warning('Topic deleted!');
+        this._router.navigate(['/home']);
+      });
+    }
   }
 
-  toggleDisplay() {
+  toggleDisplay(): void {
     this.view = (this.view === 'data') ? 'chart' : 'data';
     this.toggleButton = (this.toggleButton === 'View Chart') ? 'View Data' : 'View Chart';
   }
